requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Web upload (optional - for hosted replay parser)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
